move ClienDataStorageManager to client
