+++ libraries/stumbler/src/main/java/org/mozilla/mozstumbler/service/stumblerthread/datahandling/PersistedStats.java
@@ -42,7 +42,7 @@ public class PersistedStats {
         mContext = context.getApplicationContext();
     }
 
-    void forceBroadcastOfSyncStats() {
+    public void forceBroadcastOfSyncStats() {
         sendToListeners(readSyncStats());
     }
